# openapi.yaml
openapi: 3.1.0
info:
  title: Tutor Agent API
  version: 0.1.0
paths:
  /health:
    get:
      summary: Health
      operationId: health_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Health Health Get
  /auth/register:
    post:
      summary: Register
      operationId: register_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/login:
    post:
      summary: Login
      operationId: login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /sessions:
    get:
      summary: List Sessions
      operationId: list_sessions_sessions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSessionsResponse'
      security:
      - HTTPBearer: []
    post:
      summary: Create Session
      operationId: create_session_sessions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /sessions/{session_id}:
    get:
      summary: Get Session
      operationId: get_session_sessions__session_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete Session
      operationId: delete_session_sessions__session_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Delete Session Sessions  Session Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /sessions/{session_id}/messages:
    get:
      summary: List Messages
      operationId: list_messages_sessions__session_id__messages_get
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMessagesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /sessions/{session_id}/chat:
    post:
      summary: Chat
      operationId: chat_sessions__session_id__chat_post
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AuthResponse:
      properties:
        user:
          $ref: '#/components/schemas/UserOut'
        access_token:
          type: string
          title: Access Token
      type: object
      required:
      - user
      - access_token
      title: AuthResponse
    ChatRequest:
      properties:
        message:
          type: string
          minLength: 1
          title: Message
      type: object
      required:
      - message
      title: ChatRequest
    ChatResponse:
      properties:
        session_id:
          type: string
          format: uuid
          title: Session Id
        reply:
          $ref: '#/components/schemas/MessageOut'
      type: object
      required:
      - session_id
      - reply
      title: ChatResponse
    CreateSessionRequest:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      type: object
      title: CreateSessionRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ListMessagesResponse:
      properties:
        messages:
          items:
            $ref: '#/components/schemas/MessageOut'
          type: array
          title: Messages
      type: object
      required:
      - messages
      title: ListMessagesResponse
    ListSessionsResponse:
      properties:
        sessions:
          items:
            $ref: '#/components/schemas/SessionOut'
          type: array
          title: Sessions
      type: object
      required:
      - sessions
      title: ListSessionsResponse
    LoginRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          minLength: 1
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
    MessageOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        session_id:
          type: string
          format: uuid
          title: Session Id
        role:
          $ref: '#/components/schemas/MessageRole'
        content:
          type: string
          title: Content
        meta:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Meta
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - session_id
      - role
      - content
      - created_at
      title: MessageOut
    MessageRole:
      type: string
      enum:
      - system
      - user
      - assistant
      - tool
      title: MessageRole
    RegisterRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          minLength: 1
          title: Password
        role:
          $ref: '#/components/schemas/UserRole'
          default: student
      type: object
      required:
      - email
      - password
      title: RegisterRequest
    SessionOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - created_at
      - updated_at
      title: SessionOut
    UserOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          title: Email
        role:
          $ref: '#/components/schemas/UserRole'
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - email
      - role
      - created_at
      title: UserOut
    UserRole:
      type: string
      enum:
      - student
      - tutor
      - admin
      title: UserRole
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer